requirements.txt
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.7
django-extra-views==0.6.4
django-haystack==2.5.1
django-oscar==1.4
django-oscar-api==1.1.5
django-rest-framework==0.1.0
django-tables2==1.9.0
django-template-debug==0.3.5
django-treebeard==4.1.2
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
factory-boy==2.8.1
Faker==0.7.17
funcsigs==1.0.2
idna==2.6
ipaddress==1.0.18
mock==2.0.0
olefile==0.44
pbr==3.1.1
phonenumbers==8.5.2
Pillow==4.1.1
purl==1.3.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.4
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.21
urllib3==1.22
